add: MapフォルダーにあるComponentをui/componentに移動しました
