+++ lib/ui/screen/map/map_screen.dart
@@ -5,8 +5,8 @@ import 'package:food_gram_app/core/model/posts.dart';
 import 'package:food_gram_app/core/utils/async_value_group.dart';
 import 'package:food_gram_app/core/utils/location.dart';
 import 'package:food_gram_app/env.dart';
+import 'package:food_gram_app/ui/component/app_floating_button.dart';
 import 'package:food_gram_app/ui/component/modal_sheet/app_map_restaurant_modal_sheet.dart';
-import 'package:food_gram_app/ui/screen/map/component/map_floating_action_button.dart';
 import 'package:food_gram_app/ui/screen/map/provider/maplibre_provider.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:maplibre_gl/maplibre_gl.dart';
